Refactor url package as a standard go package practice
